code/requirements.txt
minio
s3fs
pandas
python-dotenv
fastparquet